code_upload_challenge_starters/requirements/environment.txt
grpcio==1.22.0
grpcio-tools==1.22.0
gym==0.15.4
requests==2.25.0
urllib3==1.24.2